Handle IMAP IDLE
I've now got a main loop inside of mh-gmail-watch that at least
appears to be able to deal with IMAP IDLE.  Essentially what it does
is open up a connection that does an IDLE, wait for something to come
back, refresh the IDLE, and then proceed to fetch everything over a
seperate connection.  With any luck this should be immune to the
troubles that offlineimap keeps experiencing.

This also fixes a big in line_buffer, but it's not like anyone was
using it... :).