Add mhimap::folder{,_iter}
I used to just return the folder list as a list of strings, but that's
not really type safe.  This instead returns the folder list as a list
of folder objects.  While these currently just contain strings, the
idea is that I really want these to contain a whole lot more -- the
UIDVALIDITY, for example.

In an ideal world, the folder class itself would be responsible for
taking hold of the IMAP connection for a bit and parsing the output of
a SELECT command.  I think that is going to require a bit more work,
though...