Add mhimap::message{,_iter}
This deals with messages in a much more type-safe manner.  Essentially
the idea is that messages are now a class that contain some parsed
values (currently just the UID and folder they're attached to) as
opposed to just a list of strings dumped back from the IMAP server.

I'm now approaching the ability to move messages around!