Add a MIME parser
This (at least attempts to) parse a mail message into a MIME document
tree.  It's not very well tested, but at least it's capable of parsing
some simple documents to find the text/plain part.

As a result of prior commits, "show" now prints out the text/plain
part of messages that this supports.  There aren't any character
conversions anywhere, so it doesn't work correctly with base64 encdode
text.