Have mh-gmail-watch remove messages from MH
This was actually somewhat tricky because I hadn't managed to keep the
information around in order to map IMAP messages back to MH messages.
This means I had to change the database format in order to actually
make this work.

Luckily MHng isn't actually useful for anything, so it's not like
anyone is using it... :).