Add a "purge" column to IMAP__messages
The idea here is that I don't actually need to have a transaction log,
all I actually need to do is to mark messages to purge on the server
next time a synchronization is performed.

In order to make this work I've added a column to track these sorts of
messages, which later need to be cleaned up by the IMAP
synchronization program later.