Work around a probable bug
I'm not entirely sure what's going on, but it appears that I can end
up in a state where a message exists only in the IMAP__messages table
(and not the MH__messages table).  This ends up causing trouble
because the IMAP synchronizer will attempt to remove it but is unable
to.

While this appears to indicate a bug, in this case it's not actually a
problem so I'm just going to ignore it.