Add "mh-urls", which lists the URLs in a message
Like pretty much all the MH programs, this one is designed for a
specific use case I have.  I get a lot of messages with URLs in them
(RSS feeds, for example) and I don't want to have to click on the URLs
every time.  This allows me to list the URLs in a message as well as
to open them in firefox.

Like everything else in mhng, it's pretty flaky right now.
Specifically, I'm not actually sure what the URL parsing rules are so
this probably gets them somewhat wrong.