Add "folder::canonicalize_current()", and call from folders
It turns out that I pretty frequently end up with folders that don't
have the correct sequence numbers.  This allows me to actually run the
"folders" command when I'm in that sort if situation.  I might want to
call it more often, but I'm a bit worried about this starting to cause
a whole bunch of race conditions...