Only perform a single big select for "mhng-scan"
The general idea is that I have sufficient information to only perform
a single SELECT operation, which is just about as good as it's going
to get.

I still have to perform a bunch of SELECT calls to the MH__current
table, but that's just bad database design and will eventually be
going away.