Be a bit more careful about what multpart/mixed show
The idea here is that I want to make sure to show only text-like
stuff.  To do this I just added an additional check to make sure
whatever body parts we're concatonating are actually "text/plain", as
otherwise bad things could result (we don't want an "application/pdf"
concatonated to the end of a "text/plain").