Implement "mhng-install"
This script sets up your $HOME/.mhng directory to contain what MHng
requires.  This greatly simplifies the new MHng code because I don't
have to worry about fixing up the database as it goes.

Note that this script is far from done: it should really ask the user
for a bunch of information.  That said, for now I think it's good
enough: the stuff I want to add isn't supported by the old code so
I'll need to eliminate all of that first.