Implement "mhng-pack"
This replaces the old pack code.  It's essentially the same thing,
just a lot cleaner and a lot safer.  This demonstrates how
transactions nest: libmhng uses a fairly weak transaction to move
messages around, while the pack wrapper uses a big exclusive
transaction to make sure nobody is messing with the database while
this happens (as it's a really fragile operation).