Perform my first database layout change
This removes the "MH__default" table and instead moves it to a column
within the "MH__current" table.  The general idea here is that the
current folder will be tagged with the "cur" bit, and others won't.

The big change here is that I can write constraints to enforce that
exactly one folder is actually the current folder without requiring
any sort of magic single-table entry.