Implement "mhng-comp"
This apparentlny required a whole lot of changes.  Essentially I
needed some mechanism for getting data into the database, which I
didn't do before.

Note that I also made a database change that's not backwards
compatible.  I attempted to make this work with the old stuff, but I
have no idea what's going on because it doesn't appear to work.