Implement local alias handling in mailrc
The general idea here is that I want to be able to detect addresses
that I can send from and use those addresses to populate the "From"
field in replies.  This now tracks that information (which was always
stored in the mailrc).

Like all address fields, they're not going to really be correct unless
you get them from a mailrc.  Note that this shouldn't be a problem
unless you're parsing text (either message or SQL) directly, so you
don't have to worry about it as a user of the library...