Implement "mhimap-fetch"
This uses the old IMAP library and the new MHng library to fetch
messages from the IMAP server and into folders.  It appears to work
more or less correctly, but doesn't support synchronization in the
other directions yet.

Note that I might get rid of all these and end up with just
"mhimap-sync", as it appears that I don't really go any faster by only
synchronizing in one direction...