Write the whole message, not just the body
This fails when passed a proper MIME message to insert into the
database: it only inserts the "body" part (which is the plain text
stuff in the middle), not the whole message.  The fix is to simply
insert the whole message, using the "root" part instead of the "body"
part.

I didn't notice this because before I was only inserting non-MIME
messages that I hand-wrote.