Sync the read flag to the IMAP server
Now when I read a message it'll get reflected in GMail!  This is
really great because it makes the notifications on my phone a whole
lot nicer: now they'll show the latest message as opposed to just the
first message in the thread that I haven't happened to click on yet.