Handle quoted MIME sub-headers
The general idea here is that if I've got something like

  HEADER: ... key=value key=value

then I want to split the two keys by that space, while if I've got
something more like

  HEADER: ... key="value with space" key=value

then I want to not split on those first spaces but only split on the
later space.

Note that despite the fact that the code appears to not do this, it
actually does: we're only looking for the first terminator, which in
this case will be that '"'.  At that point we're done with the parse,
so it doesn't matter that there's more spaces after...