Add "mhng-daemon", a synchronization daemon
The general idea here is that I want to have something that runs in
the background so I can pass synchronization requests intho the
background.  This is nice because it allows me to have a single point
that fires off synchronization jobs.  While this shouldn't be
necessary, it's just a bit cleaner to track everything down.

Right now this doesn't really do a whole bunch because nothing talks
to it...