Fix a IMAP fetch hang
So apparently sometimes (this is the first time it's happened) Google
will give me a message that looks like

  FETCH (...
  BODY
  BODY)

as opposed to

  FETCH (...
  BODY
  BODY
  )

like it usually does.  This used to cause me IMAP fetch handler to
barf, but I've gone ahead and added a special case for it right now.
Hopefully I didn't break anything...