Quickly retry a single IMAP failure
So the general idea here is that I have a number of operations that
cause IMAP sub-processes to fail -- for example, whenever the network
disconnects I end up killing the IDLE process so it can reliably
reconnect.

This simply allows me to recover very quickly from those failures of
this type without ending up flooding my machine if I end up without a
network connection (or flooding the server if my auth fails).