Add a thread to watch processes
I'm not 100% sure this is actually necessary, but I guess at least
it's something...  The general idea here is that waitpid() blocks and
I want to start safely killing processes (so I can deal with network
failures).  It seems that sometimes when my network goes down I just
end up with hung connections (not 100% sure why, but this happens in
lots of programs).

While I'm fairly sure that I could get around this thread with another
lock, I'm too lazy to bother -- threads are easy... :).