Don't do any network operations with a SQL lock
So it turns out that SQLite just locks everything and starts to fail
all other connections whenever locks are held.  This is a pain so I'm
just going to go ahead and just never hold any locks while going to
the network.

Note that this should all be OK because I've structured my UIDs (and
IMAP did too, good job!) in a way such that all operations are
retryable without any trouble.