Have SQLite retry for 1 second before aborting
This is a much better way to have SQLite deal with the LOCKED error.
Essentially before there was no time reserved for waiting on locks,
which make them fail all the time.  This now sets a reasonable timeout
and therefor allows me to avoid dealing with that error at all.

This is much better, because I'm not sure that retrying everything in
such a trivial manner actually produced the correct results!