Support the MHNG_NO_DAEMON enviornment variable
This is pretty much only going to be used for testing.  The idea is
that when I run tests I don't want to start forking off processes all
over the place that won't do anything.

Instead this just allows the MHng daemon to be completetly disabled.
Any requests return instantly with a dummy response.