Correctly decode base64 messages
This was a mess: the only reason I passed the tests before is because I
wasn't actually producing a .gold file to diff against.  It turns out I
was leaving some extra junk at the end of a base64 decode run because I
wasn't accounting for the extra 0-byte that the upstream baes64_decode
must be appending.