Add a newline in add_header
add_header didn't produce a newline, so as a result the next line was
eaten by the header.  If this header was the last one in line, then it
would eat the first line of the message.

I didn't notice this because most of the time I had a date and message
ID filled out by git send-email.