Allow messages to be compacted
This is a bit of a hack, where if I do something like
"folder->messages()" on a large inbox I can actually run out of memory
quite quickly just because the entire contents of the message is stored
in memory.

This allows users of a message to declare that they're done using it,
which frequently allows only a single message to be in flight at a time.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>