Allow "next" after "rmm"
This is a bit ugly.  Specifically, I've introduced the concept of a fake
message but haven't properly integrated it into the class heirarchy
because I don't want to deal with refactoring everything right now.
After letting this sit overnight it wouldn't actually be that hard: the
"maybe_fake_message" should just be a fake_message and message should be
a subtype of fake message.  I got it backwards the first time around,
though, so it didn't happen.

Either way, this works for me which is good enough for this morning.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>