Avoid crossing accounts in PURGE and FLAGS
These two synchronize lists of messages between the client and server, with
messages possibly having disappeared on the server.  Since messages that are
deleted on the server are just disappeared without a trace, we need to make
sure that the message was supposed to be in this account before disappearing it
locally.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>