Treat ks_c_5601-1987 as EUC-KR
I probably can't read any of this so it's not like I'm going to be able to tell
if this is correct, but it looks like iconv doesn't support this encoding and I
can just use a supported one instead.  I'm not going to bother testing it.

See https://marc.info/?l=kde-i18n-doc&m=100396886120761&w=2 for more info.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>