Handle multiple GNUTLS++ includes
Looks like there's some weirdness in GNUTLS where it's assuming
GNU-style inline declarations in some way that clang doesn't like.
There's some compalining about this on the inernet, but in the GUNTLS++
sources there's a #ifdef to handle this.

So all but one of the includers of gnutlsxx.h need to define the magic
macro.  Thus I just have this shim gnutlsxx.c++ that includes the header
file to get all the inline definitions.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>